change float

+++ Patch_test/test.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char* argv[])
 {
-	double a = 10;
+	float a = 10;
 	printf("a=%d\n", a);
 	return 0;
 }
